Handle optional/missing fields better
Almost all of the lap fields can be missing (even end-time!); in these
cases we should omit the resulting TCX element entirely, rather than use a
default value, and certainly we shouldn't raise an exception due to being
unable to convert None to float.

Also, some manufacturer and product IDs are not (yet) found in fitparse, and
thus only available in numeric form, which must be coerced to string.

For an example of both of the above problems, see FIT files generated by
recent versions of the Strava Android/iPhone apps.