Handle files which only have local/relative timestamps
In some files, perhaps created for non-GPS activities or by non-GPS devices,
all timestamps in laps and trackpoints are expressed as integer offsets.
(Probably an offset in seconds relative to when the device was turned on.)

The only "real" timestamp available is in the 'activity' message field
called 'local_timestamp', which represents the device's time in its *local*
timezone.

To process such files, we must use 'activity.local_timestamp' along with
'activity.timestamp' to calibrate the offset between the integer timestamps
and Unix epoch seconds.  However, even after doing this we *do not know* the
correct timezone for the resulting timestamps.  We emphasize this in the
ISO8601-format timestamps in the TCX output, by not appending 'Z'.

'tests/files/antfs-dump.63.fit' from dtcooper/python-fitparse is a good
example of this issue.  It contains data from a heart-rate monitor device
without GPS.